backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
mongoengine==0.28.2
pymongo==4.6.1
django-cors-headers==4.3.1
python-decouple==3.8